backend/requirements.txt
django>=4.2
gunicorn
dj-database-url
python-dotenv
psycopg2-binary
django-cors-headers
PyJWT
